+++ prequ/ini_parser.py
@@ -4,6 +4,8 @@ from six.moves import configparser
 
 text = type('')
 
+bool_or_auto = 'bool_or_auto'
+
 
 class ParseError(Exception):
     pass
@@ -23,12 +25,14 @@ def parse_ini(fileobj, field_specs, section_name):
 
 
 def _parse_value(parser, section_name, key, value, spec):
-    if spec == bool:
+    if spec in (bool, bool_or_auto):
+        if spec == bool_or_auto and value == 'auto':
+            return 'auto'
         try:
             return parser.getboolean(section_name, key)
         except ValueError:
             raise ParseError(
-                'Unknown value for bool option "{}": {!r}'.format(key, value))
+                'Unknown bool value for option "{}": {!r}'.format(key, value))
     elif spec == text:
         return value
     elif isinstance(spec, list):